// src/app/model/Boletim.ts
export class Boletim{
    Dia: Date;
    CalendarioLiturgico: string;

    // inicio
    ChamadaReuniao: string;
    Preludio: string;
    ChamadaAdoracao: string;
    OracaoInvocacao: string;
    CantigoCongregacional: string;

    // Confissao
    TituloConfissao: string;
    ChamadaConfissao: string;
    ResponsoConfissao: string;
    OracaoConfissao: string;
    DeclaracaoPerdao: string;

    // Louvor
    TituloLouvor: string;
    ConviteLouvorSaudacao: string;
    LouvorComunitario: string;
    OracaoCriancas: string;

    // Palavra
    TituloIluminacao: string;
    OracaoIluminacao: string;
    LeituraAntigoTestamento: string;
    ResponsoAntigoTestamento: string;
    LeituraNovoTestamento: string;
    ResponsoNovoTestamento: string;
    ProclamacaoPalavra: string;
    AfirmacaoFe: string;
    OfertorioTexto: string;
    OfertorioCantico: string;
    OracaoIntercessao: string;

    // seia
    TemSeia: boolean;

    // Envio 
    TituloEnvio: string;
    EnvioTexto: string;
    EnvioCanticoCongregacional: string;
    OracaoFinalTexto: string;
    BencaoTexto: string;
    AmemTexto: string;
    Posludio: string;
}